requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pydantic
email-validator 
python-multipart